move logic back into deploysteps
